hdcp handshake reset
Simple script that ssh's into the netv ip, and uses fpga_ctl to disconnect, and reconnect the hot plug sink to try to capture the correct hdcp key. (use this if you see rainbow static on the overlay). This happens to me every so often when switching sources via my receiver.